+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+* Add `get_bucket_name_or_none_from_bucket_id` to `AccountInfo` and `Cache`
+
+### Fixed
+* Cleanup sync errors related to directories
+* Use proper error handling in `ScanPoliciesManager`
+
+### Changed
+* `b2sdk.v1.sync` refactored to reflect `b2sdk.sync` structure
+
 ## [1.7.0] - 2021-04-22
 
 ### Added